[bug] Add a panic handler in CFFI layer
Deadlocks in the cffi layer happened due to panics in async handlers.

Now we return the fact that the panic happens in the middle of the stream.
