+++ engine/language_client_go/include/baml_cffi_generated.h
@@ -11,17 +11,17 @@
 #include <stdint.h>
 #include <stdlib.h>
 
+typedef struct Buffer {
+  const int8_t *ptr;
+  size_t len;
+} Buffer;
+
 typedef void (*CallbackFn)(uint32_t call_id,
                            int32_t is_done,
                            const int8_t *content,
                            uintptr_t length);
 
 typedef void (*OnTickCallbackFn)(uint32_t call_id);
-
-typedef struct Buffer {
-  const int8_t *ptr;
-  size_t len;
-} Buffer;
 
 const char *version(void);
 
@@ -32,10 +32,6 @@ const void *create_baml_runtime(const char *root_path,
 void destroy_baml_runtime(const void *runtime);
 
 int invoke_runtime_cli(const char *const *args);
-
-void register_callbacks(CallbackFn callback_fn,
-                        CallbackFn error_callback_fn,
-                        OnTickCallbackFn on_tick_callback_fn);
 
 /**
  * Extern "C" function that returns immediately, scheduling the async call.
@@ -62,3 +58,7 @@ struct Buffer call_object_constructor(const char *encoded_args, uintptr_t length
 void free_buffer(struct Buffer buf);
 
 struct Buffer call_object_method(const char *encoded_args, uintptr_t length);
+
+void register_callbacks(CallbackFn callback_fn,
+                        CallbackFn error_callback_fn,
+                        OnTickCallbackFn on_tick_callback_fn);
